Update OsmSharp package which has a bug with wrong IDs in current version

+++ Core/Primitives/OsmElement.cs
@@ -23,7 +23,7 @@ public abstract class OsmElement : IChunkerItem
     public long Id { get; }
 
     [PublicAPI]
-    public int Version { get; }
+    public long Version { get; }
     
     [PublicAPI]
     public long Changeset { get; }
